changed back to unsigned shorts from the erroneous unsigned ints, fixed point class to come

+++ Detectors/TRD/base/include/TRDBase/LocalGainFactor.h
@@ -26,10 +26,10 @@ namespace o2
 namespace trd
 {
 
-class LocalGainFactor : public PadCalibrations<unsigned int>
+class LocalGainFactor : public PadCalibrations<unsigned short>
 {
  public:
-  using PadCalibrations<unsigned int>::PadCalibrations;
+  using PadCalibrations<unsigned short>::PadCalibrations;
   ~LocalGainFactor() = default;
 };
 } // namespace trd
